refactor(cardano): introduce Signer and its subclasses
